+++ test/DbTestCase.php
@@ -322,7 +322,7 @@ abstract class DbTestCase extends TestCase
 
         $db->insertRow('UserThings', ['user_id' => $ids[3]]);
 
-        $actual = $entities->getEntitiesByIds([$ids[2], $ids[3]], ['id', 'name', 'isDisabled', 'computed', 'weight', 'thing.uid']);
+        $actual = $entities->getEntitiesByIds([$ids[2], $ids[3]], ['id', 'name', 'isDisabled', 'computed', 'thing.uid']);
 
         $expected = [
             [
@@ -330,7 +330,6 @@ abstract class DbTestCase extends TestCase
                 'name' => 'Modern user 4',
                 'isDisabled' => false,
                 'computed' => 41.0,
-                'weight' => $db->options->floatSelectedAsString ? '40' : 40.0,
                 'thing' => [
                     'uid' => $ids[3],
                 ],
@@ -340,7 +339,6 @@ abstract class DbTestCase extends TestCase
                 'name' => 'Modern user 3 modified',
                 'isDisabled' => false,
                 'computed' => 31.0,
-                'weight' => $db->options->floatSelectedAsString ? '30' : 30.0,
                 'thing' => null,
             ],
         ];